AbstractCRUD added, some restructuring of WebController and WebSocketConfig into UI. Everything still functional
