+++ src/Main_Controller/SpringApplicationConcrete.java
@@ -7,8 +7,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan("UI")
 public class SpringApplicationConcrete {
 
     public static void main(String[] args) {
