+++ src/UI/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package Main_Controller;
+package UI;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -12,7 +12,7 @@ public class WebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
 
     @Override
     public void configureMessageBroker(MessageBrokerRegistry config) {
-        config.enableSimpleBroker("/topic");
+        config.enableSimpleBroker("");
         config.setApplicationDestinationPrefixes("/app");
     }
